更新数据库出错的响应内容

+++ backend/sql/ddl/usr_reset.sql
@@ -15,7 +15,7 @@ CREATE TABLE "usr"."usr_info" (
 CREATE INDEX btree_idx_name ON "usr"."usr_info" USING btree (LOWER(name));
 
 CREATE TABLE "usr"."user_profiles" (
-    "usr_id"        INT             PRIMARY KEY REFERENCES "usr"."usr_info"(id),
+    "usr_id"        BIGINT          PRIMARY KEY REFERENCES "usr"."usr_info"(id),
     "biography"     TEXT            DEFAULT '# 默认用户说明',
     "avator"        TEXT            DEFAULT '默认头像图片路径（开发中）',
     "background"    TEXT            DEFAULT '默认背景图片路径（开发中）',
